Refactoring location controller testes
LocationControllerTest tests were moved to LocationControllerIntegrationTest.
@Alexandre Mello
@Lucas Carol
